added Google Analytics tracking code

+++ components/questions/Page/MoreDropdown/index.vue
@@ -1,4 +1,14 @@
 <template>
+	<a-popconfirm v-if="$mobx.account.isLoggedIn" v-on:submit="handleDeleteQuestion" title="Are you sure delete this question?"
+		@confirm="handleDeleteQuestion(question)"
+		ok-text="Yes"
+		cancel-text="No"
+		placement="top">
+		<a-button icon="delete" shape="circle"></a-button>
+	</a-popconfirm>
+</template>
+
+<template>
     <a-dropdown class="more-dropdown">
       <a-button shape="circle" icon="ellipsis" />
       <a-menu slot="overlay">
@@ -8,19 +18,19 @@
         <a-menu-item>
           <a @click="handleMarkAsSpam">Mark the question as spam</a>
         </a-menu-item>
+        <a-menu-item>
+
+          <a-popconfirm v-if="$mobx.account.isLoggedIn" @confirm="handleDeleteQuestion(question)" v-on:submit="handleDeleteQuestion" title="Are you sure delete this question?"
+          ok-text="Yes"
+      		cancel-text="No"
+          placement="top">
+          Delete
+          </a-popconfirm>
+        </a-menu-item>
       </a-menu>
     </a-dropdown>
 </template>
 
-<slot="actions" slot-scope="question">
-	<a-popconfirm v-if="$mobx.account.isLoggedIn" v-on:submit="handleDeleteQuestion" title="Are you sure delete this question?"
-		@confirm="handleDeleteQuestion(question)"
-		ok-text="Yes"
-		cancel-text="No"
-		placement="top">
-		<a-button icon="delete" shape="circle"></a-button>
-	</a-popconfirm>
-</slot>
 
 <script>
   import {MARK_SPAM, MARK_BLOCK_NOTIF} from '~/constants'
@@ -50,13 +60,19 @@
       },
 
       handleDeleteQuestion(question) {
+      if (this.$mobx.account.isLoggedIn) {
         this.isPending = true
         this.$http.delete('questions/' + question._id).then(() => {
           this.fetch()
         }).catch(() => {
           this.isPending = false
+          this.$router.push('/collection/newest/')
+          message.success('Success!') // add after this redirection and only own Q can be deleted
         })
+      } else {
+        this.$mobx.ui.loginModal.show('signup')
       }
+    }
     }
   }
 </script>
